CHANGEES MADE BY: ABS0LUT100N - 12/05/2026

+++ Clase 13/todolist/src/main/java/com/example/todolist/repository/CategoriaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.todolist.model.Categoria;
 
-public interface CategoriaRepository extends JpaRepository<Categoria,Long>{
+public interface CategoriaRepository extends JpaRepository<Categoria,Integer>{
 
     
 }
